Activity removed as an ontological entity
We now define our main functions at top-level as defined classes, each
of which are realized in specific processes. Activity is retained in
Clojure space as a term for the pair of a realisable entity and a
process.

defcatalyst and defbinding have been altered to reflect this change, and
some subsidiary functions have now been removed, as the patterns are
much simpler.